fin1/adaptive_market_making_implementation/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Machine learning libraries
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0
pymc3>=3.11.0
theano-pymc>=1.1.0
arch>=5.0.0

# Agent-based modeling
mesa>=0.9.0

# Financial libraries
QuantLib-Python>=1.24.0

# Data pipeline libraries
kafka-python>=2.0.0
influxdb-client>=1.24.0
apache-flink>=1.14.0
pyflink>=1.14.0

# Optimization libraries
scikit-optimize>=0.9.0

# Utility libraries
tqdm>=4.62.0
joblib>=1.1.0
pyyaml>=6.0.0

# Model deployment
onnx>=1.10.0
onnxruntime>=1.10.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Documentation
sphinx>=4.3.0
sphinx-rtd-theme>=1.0.0
